refactor: remove errors.ts as an entry point
