+++ src/index.ts
@@ -1,1 +1,6 @@
-export { transformSync } from "./transform.ts";
+export { transformSync } from "./internal/transform.ts";
+export {
+  SwcError,
+  isSwcError,
+  wrapAndReThrowSwcError,
+} from './internal/errors.ts';
