+++ src/transform-loader.ts
@@ -1,7 +1,6 @@
 import type { LoadFnOutput, LoadHookContext } from "node:module";
 import type { Options } from "../lib/wasm";
-import { isSwcError, wrapAndReThrowSwcError } from "./errors.js";
-import { transformSync } from "./index.js";
+import { transformSync, isSwcError, wrapAndReThrowSwcError } from "./index.js";
 
 export async function load(
 	url: string,
